Update Versioning.md
On line #27 the sentences "The first time you enable a bucket for versioning, objects in it are thereafter always versioned and given a unique version ID" 
is not entirely accurate.  The phrase "first time" is not accurate because versioning could be enabled for a bucket, then suspended, and then enabled again. The rules described here would apply to the second time versioning is enabled in a bucket. Additionally, the existing objects are not immediately modified (given a version id) once versioning is enabled, there has to be a future request to the object in order for it to be modified.  
Thank you for taking the time to review my pull request.